Differentiate last segment in Segment Lane Check
If the last segment length is not 100 meters, then the function will not give any error message. If the max to measurement value of the input data is less than the route max measurement value, then the function will give an error message.

Function rename from create_valid_df to copy_valid_df