Add Overlap Check to the 'measurement_check' Method
If the From Measure of the next segment is smaller than the previous segment To Measure, then there is an overlap between those two segment.

The segment which has an inverted From M and To M (From M > To M) will not be checked for gap or overlap with the previous segment, and the From Measure and the To Measure of that segment will be inverted for the next segment gap and overlap check.